i cant shoot

+++ include/ScriptGlobals/AutonSelectorFile.hpp
@@ -4,4 +4,4 @@ void default_A();
 void A_1();
 void A_2();
 void A_3();
-void AutonSelectorPrimary(const int autonType);
+void AutonSelectorPrimary(const u_int16_t autonType);
